ai-server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
opencv-python-headless==4.10.0.84
pillow==10.3.0
numpy==1.26.4
gunicorn==22.0.0
ultralytics==8.2.18
torch==2.4.1
torchvision==0.19.1